+++ src/webserver.rs
@@ -1,1 +1,3 @@
+pub mod createwallet;
+pub mod printchain;
 pub mod webserver;
